% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_differences.R
\name{calc_diffs}
\alias{calc_diffs}
\title{Calculate Differences}
\usage{
calc_diffs(noms_data)
}
\arguments{
\item{noms_data}{dataframe created by read_noms_data}
}
\value{
Dataframe with a column of differences
}
\description{
Calculate differences in H-data between current and most previous assessment
}
\examples{
calc_diffs(noms_data)
}
