% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_differences.R
\name{compare_diff}
\alias{compare_diff}
\title{Compare Differences}
\usage{
compare_diff(diffdata, var, zero_prob, mean, binwidth)
}
\arguments{
\item{diffdata}{dataframe created by calc_diffs}

\item{var}{H-data variable of interest}

\item{zero_prob}{(optional) The probability of the difference in the variable of interest being zero}

\item{mean}{(optional) The mean of the absolute value of the non-zero part of the variable of interest. Note that for the waist circumference:weight ratio, the mean of the absolute value of the non-zero part is multiplied by 1,000}

\item{binwidth}{(optional) The binwidth of the plot}
}
\value{
Histogram of simulated differences with lines indicating actual differences in H-data.
}
\description{
The purpose of this plot is to look for dramatic differences in the values of client-specific H-data values between the current and previous assessments. The plot may indicate outliers.
}
\details{
Simulates a comparison distribution and compares client-specific differences in H-data between current and most previous assessment to the simulated distribution.

If optional variables are not specified the function will use a reference value.
}
\examples{
compare_diff(diffdata, var, zero_prob, mean, binwidth)
}
