% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_distributions.R
\name{compare_dist}
\alias{compare_dist}
\title{Compare Distributions}
\usage{
compare_dist(newdata, var, mean, sd, binwidth)
}
\arguments{
\item{newdata}{Dataframe downloaded from SPARS containing all assessments that you want to check. This will usually just include intakes.}

\item{var}{H-data variable of interest}

\item{mean}{(optional) mean of the distribution of the variable of interest}

\item{sd}{(optional) standard deviation of the distribution of the variable of interest}

\item{binwidth}{(optional) The binwidth of the plot}
}
\value{
Histogram of simulated data with lines indicating client's actual H-data values.
}
\description{
The purpose of this plot is to search for outliers/errors/typos in client-specific H-data values.
}
\details{
Simulates a gamma distribution and compares client-specific H-data values to the simulated distribution.

If optional variables are not specified, the function will use a reference value.
}
\examples{
compare_dist(newdata, var, mean, sd, binwidth)
}
