% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_noms_report}
\alias{generate_noms_report}
\title{Generate Report}
\usage{
generate_noms_report(noms_data_path, prior_assessments_path,
  output_file = "noms_check_report.html", output_dir = getwd(),
  intermediates_dir = getwd(), knit_root_dir = getwd(), ...)
}
\arguments{
\item{noms_data_path}{Path to data created by read_noms_data.}

\item{prior_assessments_path}{(optional) Path to the csv file of prior assessments.}

\item{output_file}{Where you want the report to be saved.}
}
\value{
Report
}
\description{
Generate Report
}
\examples{
generate_noms_report("Desktop/new_noms_data.xls", "Desktop/prior_assessments.csv", "Desktop/report.html")
}
