% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_noms_data}
\alias{read_noms_data}
\title{Read NOMs data}
\usage{
read_noms_data(path)
}
\arguments{
\item{path}{Path to the ExcelReport.xlsx file}
}
\value{
Data frame with original data and derived fields
}
\description{
Read NOMs data extract and create derived fields
}
\examples{
read_noms_data("/path/to/folder/ExcelReport.xls")
}
