% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_prior.R
\name{save_prior_assessments}
\alias{save_prior_assessments}
\title{Save a csv file of prior assessments}
\usage{
save_prior_assessments(noms_data, path)
}
\arguments{
\item{noms_data}{Dataframe created by read_noms_data. This should be the PREVIOUS data set - NOT the dataset in which you are searching for errors.}

\item{path}{Where you want the csv file of prior assessments to be saved.}
}
\value{
csv file containing prior assessments
}
\description{
Takes a previous SPARS download (.xls file) and produces a csv file of prior assessments. This csv is used to filter out previous assessments from the plots so that you only see newly entered data.
}
\examples{
save_prior_assessments(noms_data, "Desktop/prior_assessments.csv")
}
